refactor : props 타입스크립트 파일명 변경 및 styled확장자 tsx로 변경
