+++ src/components/Common/Avatar/styledAvatar.tsx
