+++ src/components/Common/Card/CardProps.ts
