+++ src/components/Common/Dropdown/DropdownProps.ts
@@ -1,0 +1,4 @@
+export interface DropdownProps {
+  channelId: string;
+  setChannelId: (value: string) => void;
+}
