+++ src/components/Common/Pagination/styledPagination.tsx
