+++ src/components/Common/index.ts
@@ -1,0 +1,25 @@
+import Avatar from '@/components/Common/Avatar';
+import Badge from '@/components/Common/Badge';
+import Button from '@/components/Common/Button';
+import Card from '@/components/Common/Card';
+import Dropdown from '@/components/Common/Dropdown';
+import Image from '@/components/Common/Image';
+import Input from '@/components/Common/Input';
+import OnlineBadge from '@/components/Common/OnlineBadge';
+import Pagenation from '@/components/Common/Pagination';
+import ScrollBar from '@/components/Common/ScrollBar';
+import Text from '@/components/Common/Text';
+
+export {
+  Avatar,
+  Badge,
+  Button,
+  Card,
+  Dropdown,
+  Image,
+  Input,
+  OnlineBadge,
+  Pagenation,
+  ScrollBar,
+  Text,
+};
