+++ src/components/Footer/styledFooter.tsx
