--- src/components/Header/mockChannels.ts
@@ -1,8 +1,0 @@
-import { Channel } from '@/types/APIResponseTypes';
-
-export const mockChannels: Pick<Channel, 'posts' | '_id' | 'name'>[] =
-  Array.from({ length: 4 }, (_, i) => ({
-    posts: Array.from({ length: 5 }, (_, i) => `${i}`),
-    _id: `${i}`,
-    name: `Channel ${i}`,
-  }));
