--- src/components/Main/UserListCard/mockUsers.ts
@@ -1,256 +1,0 @@
-import { UserSnippet } from '@/components/Main/UserListCard/UserListCardTypes';
-
-export const mockUsers: UserSnippet[] = [
-  {
-    image: 'https://i.pravatar.cc/300',
-    isOnline: true,
-    isFollowing: false,
-    fullName: 'minsu',
-    _id: Math.random().toString() + Date.now().toString(),
-  },
-  {
-    image: 'https://i.pravatar.cc/330',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'min ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/340',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ha ju',
-  },
-  {
-    image: 'https://i.pravatar.cc/350',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'gong ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/360',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'hola',
-  },
-  {
-    image: 'https://i.pravatar.cc/370',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'shaeo',
-  },
-  {
-    image: 'https://i.pravatar.cc/380',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'aru',
-  },
-  {
-    image: 'https://i.pravatar.cc/390',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'goru',
-  },
-  {
-    image: 'https://i.pravatar.cc/400',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ma su',
-  },
-  {
-    image: 'https://i.pravatar.cc/410',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'emso',
-  },
-  {
-    image: 'https://i.pravatar.cc/420',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'mu',
-  },
-  {
-    image: 'https://i.pravatar.cc/430',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/440',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ju',
-  },
-  {
-    image: 'https://i.pravatar.cc/450',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'gong',
-  },
-  {
-    image: 'https://i.pravatar.cc/460',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'hola',
-  },
-  {
-    image: 'https://i.pravatar.cc/470',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'shaeo',
-  },
-  {
-    image: 'https://i.pravatar.cc/480',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'aru',
-  },
-  {
-    image: 'https://i.pravatar.cc/490',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'goru',
-  },
-  {
-    image: 'https://i.pravatar.cc/300',
-    isOnline: true,
-    isFollowing: false,
-    fullName: 'minsu',
-    _id: Math.random().toString() + Date.now().toString(),
-  },
-  {
-    image: 'https://i.pravatar.cc/330',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'min ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/340',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ha ju',
-  },
-  {
-    image: 'https://i.pravatar.cc/350',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'gong ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/360',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'hola',
-  },
-  {
-    image: 'https://i.pravatar.cc/370',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'shaeo',
-  },
-  {
-    image: 'https://i.pravatar.cc/380',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'aru',
-  },
-  {
-    image: 'https://i.pravatar.cc/390',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'goru',
-  },
-  {
-    image: 'https://i.pravatar.cc/400',
-    isOnline: true,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ma su',
-  },
-  {
-    image: 'https://i.pravatar.cc/410',
-    isOnline: false,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'emso',
-  },
-  {
-    image: 'https://i.pravatar.cc/420',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'mu',
-  },
-  {
-    image: 'https://i.pravatar.cc/430',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ja',
-  },
-  {
-    image: 'https://i.pravatar.cc/440',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'ju',
-  },
-  {
-    image: 'https://i.pravatar.cc/450',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'gong',
-  },
-  {
-    image: 'https://i.pravatar.cc/460',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'hola',
-  },
-  {
-    image: 'https://i.pravatar.cc/470',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'shaeo',
-  },
-  {
-    image: 'https://i.pravatar.cc/480',
-    isOnline: true,
-    isFollowing: true,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'aru',
-  },
-  {
-    image: 'https://i.pravatar.cc/490',
-    isOnline: false,
-    isFollowing: false,
-    _id: Math.random().toString() + Date.now().toString(),
-    fullName: 'goru',
-  },
-];
