+++ src/components/Main/UserSnippet/styledUserSnippet.tsx
