+++ src/components/Post/Detail/Content/styledContent.tsx
