+++ src/components/Post/Detail/Result/styledBarChart.tsx
