+++ src/components/Post/Detail/Vote/styledVote.tsx
