+++ src/components/PostCard/PostCardProps.ts
