+++ src/components/PostCard/styledPostCard.tsx
