+++ src/components/Setting/styledSetting.tsx
