+++ src/hooks/usePagination.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { PostSnippet } from '@/components/PostCard/PostCardTypes';
+import { PostSnippet } from '@/components/PostCard/PostCardProps';
 
 export const usePagination = (postList: PostSnippet[], limit: number = 12) => {
   const [page, setPage] = useState(1);
