+++ src/pages/PostCreate.tsx
@@ -6,11 +6,11 @@ import {
   TextAreaWrapper,
   StyledTextArea,
   ButtonWrapper,
-} from '@/components/Post/Edit/StyledPostEdit';
+} from '@/components/Post/Edit/styledPostEdit';
 import VotedBox from '@/components/Post/Edit/VoteEditBox';
 import { isValidatedForm } from '@/components/Post/Edit/formValidation';
 import { MESSAGE, PLACEHOLDER, FORM_SIZE } from '@/utils/constants';
-import { sendPostRequest } from '@/components/Post/Edit/Api';
+import { sendPostRequest } from '@/components/Post/Edit/api';
 import Button from '@/components/Common/Button';
 import Input from '@/components/Common/Input';
 import DropdownMenu from '@/components/Common/Dropdown';
